+++ OrangeDesignSystemDemo/OrangeDesignSystemDemo/Views/Guidelines/Spacings/SpacingPage.swift
@@ -115,6 +115,7 @@ struct SpacingPage: View {
         }
         .navigationTitle("Spacings")
         .navigationViewStyle(.stack)
+        .navigationbarMenuForThemeSelection()
     }
 }